+++ beam/beam-core/src/main/java/org/esa/beam/framework/dataop/dem/ElevationModel.java
@@ -68,7 +68,7 @@ public interface ElevationModel {
      * @return  an elevation in meters, or the special value returned by {@link ElevationModelDescriptor#getNoDataValue()} if an elevation is not available
      * @throws Exception if a non-runtime error occurs, e.g I/O error
      */
-    float getSample(int x, int y) throws Exception;
+    float getSample(double x, double y) throws Exception;
 
     /**
      * @return The resampling method used.
